group and separated all APIs by their collections
API collection according to Moli: account, customer, sim, subscriber
